Undo Redo für alle Objekte, die die AppBar nutzen funktioniert jetzt. Command zusammengefasst als AppBarTransfrom
