--- Frontend/VIAProMa/Assets/Scripts/UndoRedoSystem/AppBarPositionCommand.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9a0b7d0fa9c02484fb053055fd7044d7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
