--- Frontend/VIAProMa/Assets/Scripts/UndoRedoSystem/AppBarScaleCommand.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4c65b10492eb1a948b3cd5af60defecc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
